+++ CHANGELOG.md
@@ -4,8 +4,9 @@ This project follows [Semantic Versioning](http://semver.org/).
 
 ## [36.15.1] 2025-03-18
 ### Fixed
-- Fixed a bug that 'coding' type with 'optionsOrType' in answerConstraint
-  in R5 questionnaire was converted to 'coding'.
+- Fixed a bug that item.type 'coding' with answerConstraint 'optionsOrType' 
+  in R5 questionnaire should not be converted to 'choice' or 'open-choice'
+  in R4. An exception should be thrown.
   
 ## [36.15.0] 2025-02-10
 ### Changed